# app/backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.PublicMessage:
    properties:
      author_id:
        type: string
      author_type:
        type: string
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_private:
        type: boolean
      ticket_id:
        type: string
      user_info:
        $ref: '#/definitions/service.MessageUserInfo'
    type: object
...
  service.AddPublicMessageRequest:
    properties:
      body:
        type: string
    required:
    - body
    type: object
...
  service.MessageUserInfo:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Ticket Management System API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TMS API
  version: "1.0"
paths:
...
  /api/public/generate-magic-link:
    post:
      consumes:
      - application/json
      description: Generate a magic link token for testing purposes - should be removed
        in production
      parameters:
      - description: Ticket ID and Customer ID
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Magic link token and URL
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - Invalid input data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Generate magic link (Testing only)
      tags:
      - Public API
...
  /api/public/tickets/{ticketId}:
    get:
      consumes:
      - application/json
      description: Retrieve ticket details and messages using ticket ID (no authentication
        required)
      parameters:
      - description: Ticket ID
        format: uuid
        in: path
        name: ticketId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket details with messages
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - Invalid ticket ID format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get ticket by ID
      tags:
      - Public API
  /api/public/tickets/{ticketId}/messages:
    get:
      consumes:
      - application/json
      description: Retrieve paginated messages for a ticket using ticket ID
      parameters:
      - description: Ticket ID
        format: uuid
        in: path
        name: ticketId
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 50
        description: Number of messages per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of messages
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - Invalid ticket ID format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get ticket messages by ID
      tags:
      - Public API
    post:
      consumes:
      - application/json
      description: Add a new message to a ticket using ticket ID
      parameters:
      - description: Ticket ID
        format: uuid
        in: path
        name: ticketId
        required: true
        type: string
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/service.AddPublicMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Message created successfully
          schema:
            $ref: '#/definitions/handlers.PublicMessage'
        "400":
          description: Bad request - Invalid input data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ticket not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Add message by ticket ID
      tags:
      - Public API
  /api/public/tickets/tokens/{token}:
    get:
      consumes:
      - application/json
      description: Retrieve ticket details and messages using a magic link token (no
        authentication required)
      parameters:
      - description: Magic link token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ticket details with messages
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - Missing or invalid token
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or expired token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get ticket by magic link
      tags:
      - Public API
  /api/public/tickets/tokens/{token}/messages:
    get:
      consumes:
      - application/json
      description: Retrieve paginated messages for a ticket using a magic link token
      parameters:
      - description: Magic link token
        in: path
        name: token
        required: true
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 50
        description: Number of messages per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of messages
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - Missing or invalid token
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or expired token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get ticket messages by magic link
      tags:
      - Public API
    post:
      consumes:
      - application/json
      description: Add a new message to a ticket using a magic link token
      parameters:
      - description: Magic link token
        in: path
        name: token
        required: true
        type: string
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/service.AddPublicMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Message created successfully
          schema:
            $ref: '#/definitions/handlers.PublicMessage'
        "400":
          description: Bad request - Invalid input data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or expired token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Add message by magic link
      tags:
      - Public API
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"